gdk-pixbuf: fix unusable postinst when updating with mesa
As Mesa has its own libraries stored in /usr/lib/mesa for usage with
update-alternatives, this had rendered gdk-pixbuf-query-loaders
unusable when updating with `mesa` (when the package is not yet
"configured" by the package manager).

Now specifies LD_LIBRARY_PATH=/usr/lib/mesa for postinst runtime to
prevent the issue.