grub: strip all strings before kernel version number
Patch 10008 is intended to fix problems with detecting the corresponding
initrd of kernels with local names other than "aosc-main". However, the
regexp parser for "$version" still keep strings containing numeric
characters. For example:

  vmlinuz-aosc-ck5-4.7.6

will have a `$version` of

  ck5-4.7.6

because of the "5" in "ck5", which will not produce a correct pattern
when detecting initramfs:

  initramfs-ck5-4.7.6-aosc-ck5.img

This PR fixes this problem by stripping out *all* characters before
the *real* version number.