linux-kernel{,-lts}: fix container/chroot probing
Previously in postinst and prerm, when `stat -c %i /` equals to 2, it
considers the root directory not a "fake root". This is true on ext4
filesystem; however several users with Btrfs as their root filesystem have
reported that their inode number is not always 2.

Since systemd knows a lot, we use `systemd-detect-virt` to probe whether
the current environment is in a container / chroot or not.