mesa: update to 18.3.4
- Re-enable rgb10 for now, let's see what happens...
- /etc/drirc is no longer managed by the package, Mesa defaults now located
  at /usr/share/drirc.d/00-mesa-defaults.conf.

Additional patches, included with Arch Linux's package (but not Fedora's).

[PATCH] util: Get program name based on path when possible.

Some programs start with the path and command line arguments in
argv[0] (program_invocation_name). Chromium is an example of
an application using mesa that does this.

This tries to query the real path for the symbolic link /proc/self/exe
to find the program name instead. It only uses the realpath if it
was a prefix of the invocation to avoid breaking wine programs.