sed, libpipeline: disable reconfigure
The bootstrap script shipped with these packages will try to use git to
clone a copy of gnulib, and insert it into the source tree. This can
behave badly under bad network situation (or fail under network-less
situation), and this introduces extra inreproducibility to the build
result.

Disable reconfiguring these packages, as they do not really need
reconfiguring. When reconfiguring is really needed, maybe doing it
manually with autoreconf will be better to use the git-equipped
bootstrap script.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>