argyllcms: disable LTO and slightly change build script
The thread number is passed to jam w/o -j, which makes it considers the
number as a target name, and then fail. As a hack, " || true" is added,
which makes the command line more wrong. This is fixed now.

In addition, some source file cannot suffer LTO, so LTO is disabled.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>