fontconfig{,+32}: rebase AOSC modifications on top of 2.13.92
Previously, when we update to 2.13.92, we dropped some modifications to
boost font matching speed. However, these modifications is proven to be
easily rebased on top of 2.13.92 (because upstream changed nothing on
this area).

As this heavily affects performace of LibreOffice (especially when
using Chinese as LO's UI language), re-introduce these modifications,
otherwise LO is nearly unusable with Chinese UI.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>