zoom: drop due to incompatible terms of service
> Limitations on Use. You may not reproduce, resell, or distribute the Services
or any reports or data generated by the Services for any purpose unless You
have been specifically permitted to do so under a separate agreement with Zoom.
You may not offer or enable any third parties to use the Services purchased by
You, display on any website or otherwise publish the Services or any Content
obtained from a Service (other than Content created by You) or otherwise
generate income from the Services or use the Services for the development,
production or marketing of a service or product substantially similar to the
Services.