atm: "update" to 0.3.1~alpha.1
This is actually a backtrack version to make ATM at least functional in
enabling topics and installing updates automatically - though not very
reliably or cleanly, as @liushuyu cautioned.

We will have to return to libsolv one day, but as of now, we are a bit
burnt out with this whole package manager ordeal.