libvpx: update to 1.10.0
Updated for tg_owt.

Newest version is 1.11.0, but it has soname breakage, so only update to
1.10.0 now.

An ABI number change is reverted because that ABI doesn't exist in any
publicly released versions before 1.10.0, so this shouldn't be an ABI
change when taking only released versions into consideration.

This patch should be dropped when we update to 1.11.0.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>