vboot-utils: new, 110.15728.B
This version is selected because it's the last one that could be built
with a tagged flashrom version (1.3).

As the upstream uses only branches, no tags, the commit hash is taken
from the latest commit of release-R110-15278.B upstream branch.

In addition, this package suffers from the same arm64 LTO crashing issue
as flashrom.

Signed-off-by: Icenowy Zheng <uwu@icenowy.me>