core: demote some packages from Core
Demote a few packages that are not necessarily "frozen" during year-long cycles
out of Core. Keeping only Linux API Headers (linux+api), Glibc, GCC, and their
dependencies in.

Those that were moved out now merges with the rest of the tree and may be
updated at will.

Note: Kaboom will have to adapt to this change.