qemu: update to 8.1.0
- Merge in SeaBIOS and EDK2, as Qemu builds these firmware during build-time.
- Survey and specify options explicitly, where applicable.
- Add sndio, ndctl, lzfse, rdma-core, libu2f-emu, canokey-qemu, fuse-3,
  and libblkio deps.
- Drop upstreamed patch 1001.
- Rename option --enable-jemalloc => --enable-malloc=jemalloc.
- Rebase patches 1002, 1004, 1005, and 9999.

Note: --disable-rbd, we do not maintain Ceph.
Note: --disable-glusterfs, we do not maintain GlusterFS.
Note: --disable-selinux, we do not maintain SELinux (yet?).
Note: --disable-libpmem, not ready to maintain that stack yet (libfabric => pmdk).
Note: --disable-cfi, not compatible with many features.
Note: --disable-netmap, not sure how useful it is and it seems difficult to maintain.

Ref: Patch 1001 has been upstreamed at
https://github.com/qemu/qemu/commit/abe45a859b897736d7f428f58d1e5cab4fec4ddf.

Co-authored-by: Mingcong Bai <jeffbai@aosc.xyz>