liplist: update to 2.3.0
x The job requires this package to be updated and ftbfs a bit whether it is
updated or not. The source download path was changed because the old source
download path didn't seem to be labeled for the package, resulting in an
error at build time "configure: error : PACKAGE_VERSION is not defined.
Make sure to cnfigure a source tree out from git or that .tarball-version
is present"