cmake: fix postinst during intra-minor upgrades
The previous change to the postinst breaks when the user attempts to upgrade
cmake from the same minor series (i.e., 3.29.1 => 3.29.2), as the symlinked
/usr/share/cmake-x.y never gets removed (since it was already a symlink).

This revision only tests for the existence of such a matching name and removes
the symlink (as well as real directories) when found, allowing upgrades to
commence smoothly.

Also, create symlinks with relative targets.