perl: append PATH unconditionally
Previously, the profile script only appends Perl core, site, and vendor
executable paths only if the paths exist. This can cause issues where a
package creates one of those paths, as the user would not immediately be able
to make use of the executables they installed (until the next login).

Append PATHs unconditionally since it would not cause problems if any of the
paths do not exist.