replaysorcery: drop, orphaned
Originally packaged for GOverlay, but since GOverlay 1.1.1, this package
is no longer required for GOverlay to work. Moreover, the upstream has
not been active for three years, and considering that the main feature
of the software can be easily replaced with OBS Studio or other similar
video recording software, it is quite safe to orphan this package.

Signed-off-by: Kaiyang Wu <origincode@aosc.io>