openssh: guard against empty invalid or empty keys
Old systems with /etc/ssh/sshd_host_dsa_key{,.pub} would trigger re-generation
unconditionally. This is a bug - these keys were no longer recognised by
ssh-keygen, but it would return an empty output (as opposed to any sort of
error message) in stdin, giving $_key_fingerprint an empty value.

This empty value would then return a true value as it was grep'd against the
revoked lists, causing the postinst script to think that the system contained
a vulnerable key when it really didn't.

Check for empty values to prevent this.