alsa-plugins: disable maemo plugins
We do not have the runtime daemons packaged for these plugins, and the
daemons are even not available for Maemo Leste (I doubt whether these
are open source at all). In addition, the plugins are badly implemented
that it will easily crash when the expected criteria is not met, which
will get triggered when PortAudio tries to scan all audio devices.

Disable these plugins to fix (at least) Audacity startup.

Signed-off-by: Icenowy Zheng <uwu@icenowy.me>