gcc-13: new, 13.3.0
From now on we have to maintain older versions of GCC other than the
current version that is part of the AOSC OS Core, due to several
considerations and factors, including:

* CUDA requires GCC 13, due to GCC 14 switching to new standards.
* Older software requires GCC 13, for the same or similar reason.
* User requested that they need an older version of GCC. We might have
  to maintain a build of GCC 12, too.

The core-devel/gcc itself is not changed, also because of several
factors:

* We have to install a version of GCC runtime directly to /usr/lib.
* We can not risk a dependency problem during bootstrapping.
* For convenience of developers and (core package) maintainers.

The gcc package still contains the selected major release for the
current AOSC OS Core (we might want to add additional symlinks with
version suffixes, e.g. gcc-14 -> gcc). Once GCC is upgraded to a new
major release, the previous version will be packaged the same way.

- Rebase patches from aosc/releases/gcc-13.2.0.
- Use most part of prepare/beyond script from core-devel/gcc, and
  converting the sed operations into patches.
- Install GCC runtime libraries, libstdc++ and etc. to $GCC_PREFIX
  (/usr/lib/gcc/$TARGET_TRIPLE/$PKGVER) to avoid conflicts.
- Disable libiberty installation.
- Install static runtime libraries.