Revert "u-boot-menu: Enforce /boot/u-boot as the boot filesystem"
This reverts commit 442ff7f180602a06cff83e68bf6bde669263b24f.

The u-boot-menu script generates menu for extlinux, which by definition
should be able to work on an EXT partition. Even if the / is on a
filesystem that extlinux does not support, /boot could be a dedicated
EXT4 partition, which is a native supported configuration of u-boot-menu
upstream.

If the boot chain enforces a VFAT partition, putting the extlinux
configuration file (and thus the kernel) on it isn't a good idea,
because of the risk of kernel getting out-of-sync. The partition should
be mounted to /efi (or similar path) instead.

Signed-off-by: Icenowy Zheng <uwu@icenowy.me>