apr-util: drop Berkeley DB support
With GCC >= 14, Berkeley DB support no longer builds. We have been using
GDBM as the default database manager so it should not affect our
dependency tree.

Note that Berkeley DB is installed as part of BuildKit so it is always
available in the build environment. This revision disables this feature
explicitly.