cwiid: disable Python binding
The only user of cwiid is Kodi, which uses libcwiid
only.
Nobody uses this Python binding, so drop it.

And, cwiid never works on Python-3.
Although there is patch 0003 AZZRA: PyCObject_Check ->
PyCapsule_CheckExact, PyBuffer_New -> malloc which is expected to fix
the build, it never works.
malloc a new PyObject and assumes that it has the buffer interface is
absolutely buggy.

The buffer type has been removed in Python 3.0 so no matter how we
fix it, we must break the API interface, which should not be done
by a Linux distro.