desktop-base: reset nullglob to off after xdg-data-dirs-opt-apps.sh
This option, when enabled, would cause the asterisk/glob marker to be
ignored if no match is found. This resulted in a very confusing behaviour
with ls(1), where `ls *random-nonexistent*' would be treated as equivalent
to `ls' with no parameter at all.

Make sure to unset this shell option after whatever was done in
xdg-data-dirs-opt-apps.sh.

Link: https://www.gnu.org/software/bash/manual/html_node/Filename-Expansion.html