rustc: Remove upstreamed patch
The shipped libffi-sys is updated to 3.2.0 which contains the fix.

BTW the patch was deficient from day one: the correct way to patch a
shipped crate is not modifying the checksum file.  Instead a [patch]
section should be added into Cargo.toml, and Cargo.lock should be
updated with cargo update --offline -p $crate.