loonggl: Rewrite the vga_check.sh script
Using such a script instead of doing a proper check in eglGetDisplay is
already stupid.  And to make things worse the logic of this script is
flawed, causing a lot of troubles.

Rewrite it in a much better way to work around the troubles, as we don't
have the source code to fix eglGetDisplay in libEGL_loonggpu.so.